% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeback.R
\name{writebackUI}
\alias{writebackUI}
\title{按纽生成生成器界面}
\usage{
writebackUI(
  tabTitle = "反写采购合同",
  colTitles = c("操作区域", "", "显示区域"),
  widthRates = c(8, 4, 12),
  func_left = writebackUI_left,
  func_right = writebackUI_right,
  func_bottom = writebackUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
按纽生成生成器界面
}
\examples{
writebackUI()
}
